Renamed scripts to required naming convention. Replaced reference to fixture file with new name. Removed inecessary waits. Removed comments. Added files to gitignore. Added instructions to read me file. Fixed captcha issue.
