+++ cypress/fixtures/testing-users.fixture-sample.json
@@ -1,0 +1,14 @@
+{
+  "cyUserPrimaryEmailNotVerified": {
+  "username": "userNotVerified",
+  "email": "******@orcid.org",
+  "password": "**********",
+  "oid": "****-****-****-****"
+},
+"cyUserPrimaryEmaiVerified": {
+  "name": "userPrimaryEmailVerified",
+  "email": "******@orcid.org",
+  "password": "**********",
+  "oid": "****-****-****-****"
+}
+}
