+++ cypress/integration/my-orcid/sidebar-add-emails.e2e.js
@@ -1,32 +1,37 @@
 /// <reference types="cypress" />
 
-import userData from '../../fixtures/testing_users.json'
+import userData from '../../fixtures/testing-users.fixture.json'
 
 describe('My orcid - users are able to add content to their record', async function () {
   before(() => {
     cy.visit(Cypress.env('signInURL'))
     //sign in
-    cy.signin(userData.cyUser_primaryEmaiVerified)
+    cy.signin(userData.cyUserPrimaryEmaiVerified)
     cy.wait(1000)
   })
 
-  it('User changes visibility to the primary email account', function () {
+  it('User adds secondary Email to their record', function () {
+    const addSecondaryEmail = 'qa@orcid.org'
     //click on edit pencil for Emails section
     cy.get('#emails-panel').within(($myPanel) => {
       cy.get('#edit-button').click()
     })
 
-    //set visibility to public
-    //FIX: use id instead of class so it doesnt fail if the record is already public
-    cy.get(
-      'button[class="mat-focus-indicator public-button mat-icon-button mat-button-base no-selected"]'
-    ).click()
+    cy.get('#add-link').click()
 
-    //save changes
+    cy.get('#newEmailInput1').clear().type(addSecondaryEmail)
     cy.get('#save-emails-button').click()
-    cy.wait(3000)
+    cy.wait(1000)
 
-    cy.get('app-panel-privacy').should('have.attr', 'aria-label', 'PUBLIC')
+    //verify the keyword is displayed
+    cy.get('#emails-panel')
+      .within(($section) => {
+        cy.get('[class="line"]')
+      })
+      .should('contain', addSecondaryEmail)
+
+    //clean up state
+    cy.cleanEmails()
 
     //sign out
     cy.get('app-user-menu').click()
