+++ cypress/integration/registry/new-user-registration.e2e.js
@@ -40,17 +40,18 @@ describe('Register new user', async function () {
       .check({ force: true })
       .should('be.checked')
 
-    //CAPTCHA
+    //CAPTCHA    
     // Wrap iframe body into a cypress object and perform test within there
-    cy.getIframeBody('iframe').within(() => {
+     cy.getIframeBody('iframe[title="reCAPTCHA"]').within(() => {
       cy.get('.recaptcha-checkbox-border').click()
       cy.get('#recaptcha-anchor', { timeout: 10000 }).should(
         'have.class',
         'recaptcha-checkbox-checked'
       )
     })
+
     cy.get('#step-c-register-button').click()
-    cy.wait(10000) //wait 10 secs before checking for the email
+    cy.wait(10000) //TODO: fix to replace wait 10 secs before checking for the email
 
     //use gmail api to check verification email was sent
     cy.task('readAllMessages', {
